Deactivate Modal
* Put it in its own file
* Simplified the modal state hook on App.js
