+++ apps/allocations/app/components/Modal/index.js
@@ -1,0 +1,1 @@
+export { default as Deactivate } from './Deactivate'
